Created general Measurement class, that specific measurements can inherit from
